Move Insight database package outside plugins\pump
