+++ insight/src/main/java/info/nightscout/androidaps/insight/di/InsightDatabaseModule.kt
@@ -3,8 +3,8 @@ package info.nightscout.androidaps.insight.di
 import android.content.Context
 import dagger.Module
 import dagger.Provides
-import info.nightscout.androidaps.plugins.pump.insight.database.InsightDatabase
-import info.nightscout.androidaps.plugins.pump.insight.database.InsightDatabaseDao
+import info.nightscout.androidaps.insight.database.InsightDatabase
+import info.nightscout.androidaps.insight.database.InsightDatabaseDao
 import javax.inject.Singleton
 
 @Module
